Move UI components to new structure
