+++ src/pages/search/outlets/landscape/connectivity/CurrentPAConnectivity.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import InfoIcon from "@mui/icons-material/Info";
 
-import { PointFilledLegend } from "pages/search/shared_components/CssLegends";
+import { PointFilledLegend } from "pages/search/ui/CssLegends";
 import { ShortInfo } from "core/composites/ShortInfo";
-import { IconTooltip } from "pages/search/shared_components/Tooltips";
+import { IconTooltip } from "pages/search/ui/Tooltips";
 import {
   SearchLegacyCTX,
   type LegacyContextValues,
@@ -11,7 +11,7 @@ import {
 import matchColor from "pages/search/utils/matchColor";
 import BackendAPI from "pages/search/api/backendAPI";
 import formatNumber from "pages/search/utils/format";
-import TextBoxes from "pages/search/shared_components/TextBoxes";
+import TextBoxes from "pages/search/ui/TextBoxes";
 
 import { currentPAConn, DPCKeys, DPC } from "pages/search/types/connectivity";
 import { textsObject } from "pages/search/types/texts";
