+++ src/pages/search/outlets/landscape/connectivity/TimelinePAConnectivity.tsx
@@ -2,14 +2,14 @@ import React from "react";
 import InfoIcon from "@mui/icons-material/Info";
 
 import { ShortInfo } from "core/composites/ShortInfo";
-import { IconTooltip } from "pages/search/shared_components/Tooltips";
+import { IconTooltip } from "pages/search/ui/Tooltips";
 import {
   SearchLegacyCTX,
   type LegacyContextValues,
 } from "pages/search/hooks/SearchContext";
 import matchColor from "pages/search/utils/matchColor";
 import processDataCsv from "pages/search/utils/processDataCsv";
-import TextBoxes from "pages/search/shared_components/TextBoxes";
+import TextBoxes from "pages/search/ui/TextBoxes";
 import {
   timelinePAConn,
   timeLinePAConnValues,
