+++ src/pages/search/shared_components/charts/SmallBars.tsx
@@ -6,7 +6,7 @@ import { darkenColor } from "pages/search/utils/colorUtils";
 import formatNumber from "pages/search/utils/format";
 import withMessageWrapper from "pages/search/shared_components/charts/withMessageWrapper";
 import { SmallBarTooltip } from "pages/search/types/charts";
-import { Tick } from "pages/search/shared_components/CssTicks";
+import { Tick } from "pages/search/ui/CssTicks";
 
 interface Props {
   data: Array<SmallBarsData>;
