+++ src/pages/search/ui/CssIcons.tsx
