+++ src/pages/search/ui/CssLegends.tsx
