+++ src/pages/search/ui/CssTicks.tsx
