+++ src/pages/search/ui/GradientLegend.tsx
