+++ src/pages/search/ui/TextBoxes.tsx
@@ -5,8 +5,8 @@ import CollectionsBookmarkIcon from "@mui/icons-material/CollectionsBookmark";
 import FormatQuoteIcon from "@mui/icons-material/FormatQuote";
 
 import { ShortInfo } from "core/composites/ShortInfo";
-import DownloadCSV from "pages/search/shared_components/DownloadCSV";
-import { IconTooltip } from "pages/search/shared_components/Tooltips";
+import DownloadCSV from "pages/search/ui/DownloadCSV";
+import { IconTooltip } from "pages/search/ui/Tooltips";
 
 interface TextBoxProps {
   downloadData?: Array<unknown>;
